+++ lua/obsidian/path.lua
@@ -98,7 +98,8 @@ end
 ---@field name string|? The final path component, if any.
 ---@field suffix string|? The final extension of the path, if any.
 ---@field suffixes string[] A list of all of the path's extensions.
----@field stem string|? The final path component, without its suffix.
+---@field stem string The final path component, without its suffix.
+---@operator div(string|obsidian.Path): obsidian.Path
 local Path = {}
 
 Path.__tostring = function(self)
@@ -314,7 +315,7 @@ end
 ---@return string?
 ---@private
 Path.abspath = function(self)
-  local path = vim.loop.fs_realpath(vim.fn.resolve(self.filename))
+  local path = vim.uv.fs_realpath(vim.fn.resolve(self.filename))
   return path
 end
 
@@ -438,14 +439,14 @@ Path.vault_relative_path = function(self, opts)
 
   local api = require "obsidian.api"
 
-  local ws = api.find_workspace(tostring(self))
+  local root = api.resolve_workspace_dir()
 
   -- NOTE: we don't try to resolve the `path` here because that would make the path absolute,
   -- which may result in the wrong relative path if the current working directory is not within
   -- the vault.
 
   local ok, relative_path = pcall(function()
-    return self:relative_to(ws.root)
+    return self:relative_to(root)
   end)
 
   if ok and relative_path then