+++ lua/obsidian/types.lua
@@ -7,9 +7,8 @@
 ---
 ---@field type "insert_template"
 ---@field template_name string|obsidian.Path The name or path of the template being used.
----@field template_opts obsidian.config.TemplateOpts The template options being used.
----@field templates_dir obsidian.Path The folder containing the template file.
----@field location [number, number, number, number] `{ buf, win, row, col }` location from which the request was made.
+---@field templates_dir obsidian.Path|? The folder containing the template file.
+---@field location [integer, integer, integer, integer] `{ buf, win, row, col }` location from which the request was made.
 ---@field partial_note? obsidian.Note An optional note with fields to copy from.
 
 ---@class obsidian.CloneTemplateContext
@@ -17,20 +16,12 @@
 ---
 ---@field type "clone_template"
 ---@field template_name string|obsidian.Path The name or path of the template being used.
----@field template_opts obsidian.config.TemplateOpts The template options being used.
----@field templates_dir obsidian.Path The folder containing the template file.
+---@field templates_dir obsidian.Path|? The folder containing the template file.
 ---@field destination_path obsidian.Path The path the cloned template will be written to.
 ---@field partial_note obsidian.Note The note being written.
 
 ---@alias obsidian.TemplateContext obsidian.InsertTemplateContext | obsidian.CloneTemplateContext
 ---The table passed to user substitution functions. Use `ctx.type` to distinguish between the different kinds.
-
----@class obsidian.workspace.WorkspaceSpec
----
----@field path string|(fun(): string)|obsidian.Path|(fun(): obsidian.Path)
----@field name string|?
----@field strict boolean|? If true, the workspace root will be fixed to 'path' instead of the vault root (if different).
----@field overrides obsidian.config?
 
 ---@class obsidian.config
 ---@field workspaces obsidian.workspace.WorkspaceSpec[]
@@ -43,17 +34,13 @@
 ---@field wiki_link_func? fun(opts: {path: string, label: string, id: string|?}): string
 ---@field markdown_link_func? fun(opts: {path: string, label: string, id: string|?}): string
 ---@field preferred_link_style? obsidian.config.LinkStyle
----@field follow_url_func? fun(url: string)
----@field follow_img_func? fun(img: string)
----@field note_frontmatter_func? (fun(note: obsidian.Note): table)
----@field disable_frontmatter? (fun(fname: string?): boolean)|boolean
 ---@field frontmatter? obsidian.config.FrontmatterOpts
 ---@field backlinks? obsidian.config.BacklinkOpts
 ---@field completion? obsidian.config.CompletionOpts
 ---@field picker? obsidian.config.PickerOpts
 ---@field daily_notes? obsidian.config.DailyNotesOpts
 ---@field open_notes_in? obsidian.config.OpenStrategy
----@field ui? obsidian.config.UIOpts | table<string, any>
+---@field ui? obsidian.config.UIOpts
 ---@field attachments? obsidian.config.AttachmentsOpts
 ---@field callbacks? obsidian.config.CallbackConfig
 ---@field legacy_commands? boolean
@@ -63,10 +50,10 @@
 ---@field checkbox? obsidian.config.CheckboxOpts
 ---@field comment? obsidian.config.CommentOpts
 ---@field search? obsidian.config.SearchOpts
+---@field note? obsidian.config.NoteOpts
 
 ---@class obsidian.config.Internal
----@field dir string|?
----@field workspaces obsidian.workspace.WorkspaceSpec[]|?
+---@field workspaces obsidian.workspace.WorkspaceSpec[]
 ---@field log_level integer
 ---@field notes_subdir string|?
 ---@field templates obsidian.config.TemplateOpts
@@ -76,15 +63,13 @@
 ---@field wiki_link_func (fun(opts: {path: string, label: string, id: string|?}): string)
 ---@field markdown_link_func (fun(opts: {path: string, label: string, id: string|?}): string)
 ---@field preferred_link_style obsidian.config.LinkStyle
----@field follow_url_func fun(url: string)|?
----@field follow_img_func fun(img: string)|?
 ---@field frontmatter obsidian.config.FrontmatterOpts
 ---@field backlinks obsidian.config.BacklinkOpts
 ---@field completion obsidian.config.CompletionOpts
 ---@field picker obsidian.config.PickerOpts
 ---@field daily_notes obsidian.config.DailyNotesOpts
 ---@field open_notes_in obsidian.config.OpenStrategy
----@field ui obsidian.config.UIOpts | table<string, any>
+---@field ui obsidian.config.UIOpts
 ---@field attachments obsidian.config.AttachmentsOpts
 ---@field callbacks obsidian.config.CallbackConfig
 ---@field legacy_commands boolean
@@ -94,3 +79,4 @@
 ---@field checkbox obsidian.config.CheckboxOpts
 ---@field comment obsidian.config.CommentOpts
 ---@field search obsidian.config.SearchOpts
+---@field note obsidian.config.NoteOpts